tasj

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,9 +6,9 @@
  */
 int main()
 {
-	int a = 0;
-	int b = 1;
-	int sum = 0;
+	long int a = 0;
+	long int b = 1;
+	long int sum = 0;
 	int i = 0;
 
 	for (i = 0; i <= 50; i++){
